+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Bug #57: Update documentation to use angle brackets for `<circle>` and `<g>` element references (@terabytesoftw)
 - Enh #58: Add `Filter` class and corresponding tests for SVG `<filter>` element functionality (@terabytesoftw)
 - Bug #59: Add exception for invalid `clipPathUnits` value in `clipPathUnits` method of `ClipPath` class (@terabytesoftw)
+- Enh #60: Add `ForeignObject` class and corresponding tests for SVG `<foreignObject>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024